Arduino Code
The code monitors the temp, Humidity, Soil moisture, and dependent on the temp turns the lights on or off to keep the germinator at a set temp. The soil moisture monitor setting allows for the pump to be turned on or off to keep the seedlings moist. The results are linked to Blynk. I have added two Blynk buttons for the lamp and pump. Still need to perfect these as they interfere with the normal temp / soil moisture monitoring.